add a content wrapper to user's my order view
